[CL-694] improved layout sanitization specs (split into controller and service specs) + updated changelog
